requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.6

# Configuration
python-dotenv==1.0.1

# Data Validation - Versions with pre-built wheels
# Note: If using Python 3.14, consider using Python 3.11 or 3.12 for better compatibility
pydantic==2.8.2
pydantic-settings==2.4.0

# AI/ML Services
openai>=1.40.0
weaviate-client>=4.0.0

# Text Processing
beautifulsoup4==4.12.3
pdfplumber==0.10.4
python-docx>=1.1.2
pymupdf==1.23.8          # PyMuPDF for PDF processing with bounding boxes
pillow==10.1.0           # Image processing for cropping and highlighting
aiofiles==23.2.1         # Async file operations

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
